requirements.txt
customtkinter
pandas
fpdf2
pillow
matplotlib
psycopg2-binary
